add unit tests for madgwick and begin integration testing against 6dof vn-100 test dataset

+++ include/interfaces/i_imu_data_source.h
@@ -5,7 +5,7 @@
 
 namespace AHRS {
   struct IMUReading {
-    double timestamp;
+    float timestamp;
     Eigen::Vector3f gyro;
     Eigen::Vector3f accel;
     Eigen::Vector3f mag;
